+++ packages/contracts/deploy/40_finalize-managing-dao/40_revoke-permissions.ts
@@ -113,4 +113,4 @@ const func: DeployFunction = async function (hre: EHRE) {
   );
 };
 export default func;
-func.tags = ['RegisterManagingDAO'];
+func.tags = ['RevokeManagingPermissionsDAO'];